requirements.txt
Flask==3.1.2
gunicorn==23.0.0
requests==2.32.5
google-auth-oauthlib==1.2.0
google-api-python-client==2.104.0
google-auth==2.41.1
google-api-core
google-auth
